+++ R/bayesian.R
@@ -56,8 +56,8 @@
 #' bayesian(mode = "regression")
 #' \dontrun{
 #' bayesian_mod <-
-#'   bayesian() %>%
-#'   set_engine("brms") %>%
+#'   bayesian() |>
+#'   set_engine("brms") |>
 #'   fit(
 #'     rating ~ treat + period + carry + (1 | subject),
 #'     data = inhaler
@@ -77,7 +77,7 @@ bayesian <-
            knots = NULL,
            stanvars = NULL,
            fit = NULL,
-           inits = NULL,
+           init = NULL,
            chains = NULL,
            iter = NULL,
            warmup = NULL,
@@ -104,7 +104,7 @@ bayesian <-
       knots = rlang::enquo(knots),
       stanvars = rlang::enquo(stanvars),
       fit = rlang::enquo(fit),
-      inits = rlang::enquo(inits),
+      init = rlang::enquo(init),
       chains = rlang::enquo(chains),
       iter = rlang::enquo(iter),
       warmup = rlang::enquo(warmup),
@@ -176,10 +176,10 @@ translate.bayesian <- function(x, engine = x$engine, ...) {
 #'
 #' @examples
 #'
-#' model <- bayesian(inits = "random")
+#' model <- bayesian(init = "random")
 #' model
-#' update(model, inits = "0")
-#' update(model, inits = "0", fresh = TRUE)
+#' update(model, init = "0")
+#' update(model, init = "0", fresh = TRUE)
 #' @method update bayesian
 #' @rdname bayesian
 #' @export
@@ -193,7 +193,7 @@ update.bayesian <-
            knots = NULL,
            stanvars = NULL,
            fit = NULL,
-           inits = NULL,
+           init = NULL,
            chains = NULL,
            iter = NULL,
            warmup = NULL,
@@ -214,12 +214,6 @@ update.bayesian <-
            silent = NULL,
            fresh = FALSE,
            ...) {
-    parsnip::update_dot_check(...)
-
-    if (!is.null(parameters)) {
-      parameters <- parsnip::check_final_param(parameters)
-    }
-
     args <- list(
       formula.override = rlang::enquo(formula.override),
       family = rlang::enquo(family),
@@ -228,7 +222,7 @@ update.bayesian <-
       knots = rlang::enquo(knots),
       stanvars = rlang::enquo(stanvars),
       fit = rlang::enquo(fit),
-      inits = rlang::enquo(inits),
+      init = rlang::enquo(init),
       chains = rlang::enquo(chains),
       iter = rlang::enquo(iter),
       warmup = rlang::enquo(warmup),
@@ -249,33 +243,13 @@ update.bayesian <-
       silent = rlang::enquo(silent)
     )
 
-    args <- parsnip::update_main_parameters(args, parameters)
-
-    eng_args <- parsnip::update_engine_parameters(object$eng_args, ...)
-
-    if (fresh) {
-      object$args <- args
-      object$eng_args <- eng_args
-    } else {
-      null_args <- purrr::map_lgl(args, parsnip::null_value)
-      if (any(null_args)) {
-        args <- args[!null_args]
-      }
-      if (length(args) > 0) {
-        object$args[names(args)] <- args
-      }
-      if (length(eng_args) > 0) {
-        object$eng_args[names(eng_args)] <- eng_args
-      }
-    }
-
-    parsnip::new_model_spec(
-      "bayesian",
-      args = object$args,
-      eng_args = object$eng_args,
-      mode = object$mode,
-      method = NULL,
-      engine = object$engine
+    parsnip::update_spec(
+      object = object,
+      parameters = parameters,
+      args_enquo_list = args,
+      fresh = fresh,
+      cls = "bayesian",
+      ...
     )
   }