+++ NEWS.md
@@ -1,3 +1,13 @@
+# bayesian 0.0.9
+
+- Depends on `parsnip >= 1.0.0`
+- Using base R pipe and depends on R >= 4.1.0
+- Supported the use of case weights
+- Supported threshold (inverse probability weights) for multi-class predictions
+- Fixed inconsistent probability (`type = "prob"`) predictions for two-class models
+- Updated package dependencies
+- Used `init` instead of the deprecated `inits` (`brms >= 2.16.9`)
+
 # bayesian 0.0.8
 
 - Updated package website